[global] Reduce the number of support nodes in a knet from 64K to 256
realistically speaking nobody is ever going to deploy more than 32/64 nodes.

NOTE: this commit changes both onwire and API!

+++ kronosnetd/cfg.h
@@ -21,7 +21,7 @@
 struct knet_cfg_eth {
 	tap_t tap;
 	int auto_mtu;
-	uint16_t node_id;
+	uint8_t node_id;
 };
 
 struct knet_cfg_ring {
