+++ kronosnetd/etherfilter.c
@@ -37,10 +37,10 @@ int ether_host_filter_fn (void *private_data,
 			  const unsigned char *outdata,
 			  ssize_t outdata_len,
 			  uint8_t tx_rx,
-			  uint16_t this_host_id,
-			  uint16_t src_host_id,
+			  uint8_t this_host_id,
+			  uint8_t src_host_id,
 			  int8_t *channel,
-			  uint16_t *dst_host_ids,
+			  uint8_t *dst_host_ids,
 			  size_t *dst_host_ids_entries)
 {
 	struct ether_header *eth_h = (struct ether_header *)outdata;
