+++ libknet/onwire.h
@@ -33,14 +33,14 @@ struct knet_hinfo_link {
 } __attribute__((packed));
 
 struct knet_hinfo_link_table {
-	uint16_t		khlt_node_id;
+	uint8_t			khlt_node_id;
 	uint8_t			khlt_local; /* we have this node connected locally */
 	struct knet_hinfo_link	khlt_link[KNET_MAX_LINK]; /* info we send about each link in the node */
 } __attribute__((packed));
 
 struct link_table {
-	uint16_t	khdt_host_entries;
-	uint8_t		khdt_host_maps[0]; /* array of knet_hinfo_link_table[khdt_host_entries] */
+	uint8_t			khdt_host_entries;
+	uint8_t			khdt_host_maps[0]; /* array of knet_hinfo_link_table[khdt_host_entries] */
 } __attribute__((packed));
 #endif
 
@@ -78,7 +78,7 @@ union knet_hostinfo_payload {
 struct knet_hostinfo {
 	uint8_t				khi_type;	/* type of hostinfo we are sending */
 	uint8_t				khi_bcast;	/* hostinfo destination bcast/ucast */
-	uint16_t			khi_dst_node_id;/* used only if in ucast mode */
+	uint8_t				khi_dst_node_id;/* used only if in ucast mode */
 	union knet_hostinfo_payload	khi_payload;
 } __attribute__((packed));
 
@@ -159,7 +159,7 @@ union knet_header_payload {
 struct knet_header {
 	uint8_t				kh_version; /* pckt format/version */
 	uint8_t				kh_type;    /* from above defines. Tells what kind of pckt it is */
-	uint16_t			kh_node;    /* host id of the source host for this pckt */
+	uint8_t				kh_node;    /* host id of the source host for this pckt */
 	uint8_t				kh_pad1;    /* make sure to have space in the header to grow features */
 	uint8_t				kh_pad2;
 	union knet_header_payload	kh_payload; /* union of potential data struct based on kh_type */
