+++ libknet/tests/api_knet_handle_enable_filter.c
@@ -25,10 +25,10 @@ static int dhost_filter(void *pvt_data,
 			const unsigned char *outdata,
 			ssize_t outdata_len,
 			uint8_t tx_rx,
-			uint16_t this_host_id,
-			uint16_t src_host_id,
+			uint8_t this_host_id,
+			uint8_t src_host_id,
 			int8_t *dst_channel,
-			uint16_t *dst_host_ids,
+			uint8_t *dst_host_ids,
 			size_t *dst_host_ids_entries)
 {
 	return 0;
