+++ libknet/tests/api_knet_host_get_host_list.c
@@ -22,7 +22,7 @@ static void test(void)
 {
 	knet_handle_t knet_h;
 	int logfds[2];
-	uint16_t host_ids[KNET_MAX_HOST];
+	uint8_t host_ids[KNET_MAX_HOST];
 	size_t host_ids_entries;
 
 	printf("Test knet_host_get_host_list incorrect knet_h\n");
